requirements.txt
fastapi==0.115.6
python-telegram-bot==21.7
pydantic=2.10.4